Convert long frame to wide frame
This PR adds the support to convert a Long formatted time series
Frame to a Wide format. The main functionality is done by calling
function LongToWide() from "github.com/grafana/grafana-plugin-sdk-go/data"
package. For the full description of this function, check the
function description in the package source code.

One of the major purpose of this PR is to adding field labels to
frame, which is required when apply "Labels to fields" transformation.

Fixes: #15